examples/langchain-python/requirements.txt
# SPDX-FileCopyrightText: 2025 Deutsche Telekom AG and others
#
# SPDX-License-Identifier: Apache-2.0

langchain
langchain-openai
python-dotenv
